dependencies: DependencyResolver now throws errors
Instead of tasking the dependency resolver with both performing it's intended
functions, as well as handling errors and showing the application modal, we are
now just asking it to throw an exception which will be handled by
ViewerDependencies. This class will be responsible for handling the exceptions
and displaying the error modal with the correct content.
